Move Branding to a more appropriate place
